Split esil into esil_ops
