+++ libr/include/r_esil.h
@@ -314,6 +314,7 @@ R_API void r_esil_free(REsil *esil);
 R_API bool r_esil_runword(REsil *esil, const char *word);
 R_API bool r_esil_parse(REsil *esil, const char *str);
 R_API bool r_esil_dumpstack(REsil *esil);
+// XXX must be internal imho
 R_API bool r_esil_mem_read(REsil *esil, ut64 addr, ut8 *buf, int len);
 R_API bool r_esil_mem_read_silent(REsil *esil, ut64 addr, ut8 *buf, int len);
 R_API bool r_esil_mem_write(REsil *esil, ut64 addr, const ut8 *buf, int len);
@@ -353,9 +354,12 @@ R_API bool r_esil_set_op(REsil *esil, const char *op, REsilOpCb code, ut32 push,
 R_API REsilOp *r_esil_get_op(REsil *esil, const char *op);
 R_API void r_esil_del_op(REsil *esil, const char *op);
 R_API void r_esil_stack_free(REsil *esil);
+R_API int r_esil_condition(REsil *esil, const char *str);
+
+// R2_600 - unify these 3 functions into one
 R_API int r_esil_get_parm_type(REsil *esil, const char *str);
 R_API int r_esil_get_parm(REsil *esil, const char *str, ut64 *num);
-R_API int r_esil_condition(REsil *esil, const char *str);
+R_API bool r_esil_get_parm_size(REsil *esil, const char *str, ut64 *num, int *size);
 
 // esil_handler.c
 R_API bool r_esil_handlers_init(REsil *esil);
@@ -413,6 +417,9 @@ R_API void r_esil_trace_show(REsil *esil, int idx, int format);
 R_API void r_esil_trace_restore(REsil *esil, int idx);
 R_API ut64 r_esil_trace_loopcount(REsilTrace *etrace, ut64 addr);
 R_API void r_esil_trace_loopcount_increment(REsilTrace *etrace, ut64 addr);
+
+R_API bool r_esil_reg_write_silent(REsil *esil, const char *name, ut64 num);
+R_API bool r_esil_reg_read_nocallback(REsil *esil, const char *regname, ut64 *num, int *size);
 
 extern REsilPlugin r_esil_plugin_null;
 extern REsilPlugin r_esil_plugin_dummy;
