+++ libr/include/r_util/r_num.h
@@ -92,6 +92,7 @@ R_API ut64 r_num_get(RNum *num, const char *str);
 R_API int r_num_to_bits(char *out, ut64 num);
 R_API int r_num_to_ternary(char *out, ut64 num);
 R_API int r_num_rand(int max);
+R_API ut64 r_num_genmask(int max);
 R_API void r_num_irand(void);
 R_API ut64 r_get_input_num_value(RNum *num, const char *input_value);
 R_API bool r_is_valid_input_num_value(RNum *num, const char *input_value);
