+++ libr/util/unum.c
@@ -911,3 +911,16 @@ R_API char *r_num_list_join(RList *str, const char *sep) {
 	}
 	return r_strbuf_drain (sb);
 }
+
+/* Returns the number that has bits + 1 least significant bits set. */
+R_API ut64 r_num_genmask(int bits) {
+	ut64 m = UT64_MAX;
+	if (bits > 0 && bits < 64) {
+		m = (ut64)(((ut64)(2) << bits) - 1);
+		if (!m) {
+			m = UT64_MAX;
+		}
+	}
+	return m;
+}
+
